rank_decomposition, diagonal_solve, LU stuff
